nnet_from_scratch/requirements.txt
numpy
scikit-learn
matplotlib
python-mnist
